better layout and adding template type to dropdown
